// types/esxi/accounts.ts
export interface Accounts {
    message: string;
    datetime: string;
    timestamp_desc: "Passwd File Modified";
    artifact: "ESXi User Account";
    data_type: "esxi:accounts:entry";
    evidence: string;
    uid: number;
    gid: number;
    info: string;
    shell: string;
    home: string;
}